[refactor] 바뀐 store 이름, 파일이름 반영
