+++ FE/src/pages/Chart/index.tsx
@@ -12,7 +12,7 @@ import useDefaultPayment from '../../service/useDefaultPayment';
 import useLoginCheck from '../../service/useLoginCheck';
 
 import { useRootData } from '../../store/DateInfo/dateInfoHook';
-import { useTransactionData } from '../../store/TransactionData/transactionInfoHook';
+import { useTransactionData } from '../../store/AccountBook/accountBookInfoHook';
 import useAccountBook from '../../service/useAccountBookSetting';
 
 import './chart.scss';
