+++ FE/src/service/useAccountBookSetting.ts
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 
-import { useTransactionData } from '../store/TransactionData/transactionInfoHook';
+import { useTransactionData } from '../store/AccountBook/accountBookInfoHook';
 
 const useAccountBook = accountBookId => {
   const setAccountBook = useTransactionData(store => store.setAccountBook);
