+++ FE/src/store/AccountBook/accountBookDataContext.tsx
@@ -1,6 +1,6 @@
 import React, { createContext } from 'react';
 import { useLocalObservable } from 'mobx-react';
-import { createStore, TStore } from './transactionData';
+import { createStore, TStore } from './accountBookData';
 
 export const transactionInfoContext = createContext<TStore | null>(null);
