+++ FE/src/store/AccountBook/accountBookInfoHook.ts
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useObserver } from 'mobx-react';
-import { transactionInfoContext } from './transactionDataContext';
-import { TStore } from './transactionData';
+import { transactionInfoContext } from './accountBookDataContext';
+import { TStore } from './accountBookData';
 
 export const useStoreData = <Selection, ContextData, Store>(
   context: React.Context<ContextData>,
